+++ gems.rb
@@ -31,4 +31,3 @@ group :test do
 	gem "bake-test"
 	gem "bake-test-external"
 end
-